+++ doc/changes.md
@@ -472,7 +472,8 @@ as part of K95 at this time, the default terminal remains VT220 for now.
  - Fixed an issue in the dialer which could cause it to crash if the path was
    too long
  - The x86-64 version of Kermit 95 now works on Windows XP x64 Edition
- - Fixed DECCRA not copying attributes or obeying DECOM
+ - Fixed DECCRA not copying attributes, obeying DECOM or properly handling
+   double-width lines
  - Fixed DECXCPR response - it was leaving the '?' character out
  - Fixed DECSCPP parameter not being optional, and setting the terminal to maximum
    width if the parameters value is 0. Any value less than 80 will now produce an
@@ -530,8 +531,14 @@ as part of K95 at this time, the default terminal remains VT220 for now.
    receiving a lot of data. Previously K95 would delete and recreate the buffer,
    now instead it will just reuse the existing one erasing lines as they are
    reused.
+ - Fixed double-width lines not clearing the portion of the line that falls
+   outside the screen
+ - Reverse video attribute now applies the bold attribute to the background for
+   DEC VT emulations as this is what the real terminals do.
  - Fixed position 5/15 (0x5F) in the DEC Special Graphics Character set being
    underscore when it should have been the empty character / null.
+ - Fixed SGR font attributes not turning off if the reverse video attribute is 
+   set (K95 bug 843)
 
 ## Kermit 95 v3.0 beta 7 - 27 January 2025